Tests/LibWeb/Text/input/css/absolutized-background-size-value.html

<!DOCTYPE html>
<style>
    #foo {
        background-size: calc(1em - 1px) calc(1em - 1px);
    }
</style>
<div id="foo"></div>
<script src="../include.js"></script>
<script>
    test(() => {
        println(getComputedStyle(foo).backgroundSize);
    });
</script>
