Tests/LibWeb/Text/expected/css/absolutized-background-size-value.txt

calc(1em - 1px) calc(1em - 1px)
